Delete SidebarDesignSpace
